Change savez to savez_compressed
The numpy savez command was used in only two modules. If this is changed
to savez_compressed, then from now on the numpy objects will be saved in
a compressed format. Whether they are compressed or not, all of these
saved numpy object files will end in '*.npz' files can be read by the
numpy.load command.

Closes: issue-124